+++ CHANGES.md
@@ -7,6 +7,7 @@
   variable.
 - [IMPROVED] Updated documentation by replacing deprecated Cloudant links with
   the latest bluemix.net links.
+- [REMOVED] Remove previously deprecated method `set_permissions`.
 
 # 1.10.0 (2017-11-01)
 - [UPGRADED] Upgrade package: cloudant-nano@6.7.0.